requirements.txt
pandas
kafka-python
psycopg2
snowflake-connector-python